backend/requirements.txt
arrow==0.17.0
asgiref==3.3.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
blessed==1.17.12
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
crispy-bootstrap5==0.4
cryptography==3.4.4
decorator==4.4.1
defusedxml==0.7.0rc2
discord-webhook==0.11.0
Django==3.0.5
django-bootstrap4==2.2.0
django-chartjs==2.0.0
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-datatables-ajax @ git+https://github.com/peppelinux/django-datatables-ajax.git@515f6e3b8ff07afb136da26da4c09cb881f7c02f
django-datatables-server-side @ git+https://github.com/monnierj/django-datatables-server-side.git@dbde62d28574d4692aac2933f16b5e9e021f5083
django-debug-toolbar==3.2.1
django-filter==2.3.0
django-markdown==0.8.4
django-markdownx==3.0.1
django-paypal==1.0.0
django-picklefield==2.1.1
django-q==1.2.1
django-suit==0.2.28
django-tables2==2.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
docopt==0.6.2
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
markdown2==2.3.9
monotonic==1.5
mysql-connector==2.2.9
mysqlclient==2.0.1
oauthlib==3.1.0
paramiko==2.7.2
Pillow==8.1.0
pycparser==2.20
pycryptodomex==3.10.1
PyJWT==2.0.1
PyNaCl==1.4.0
python-a2s==1.1.4
python-dateutil==2.8.1
python-valve==0.2.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
soupsieve==2.2
sqlparse==0.4.1
steam==1.0.2
steamwebapi==0.1.3
urllib3==1.26.3
valve==0.0.0
vdf==3.3
wcwidth==0.2.5
zipp==3.4.0